Add store and list messages feature and refactor
Chatbot commands are now exclusively lower case.

This is to keep behaviour consistent in supporting future commands.

Looping behaviour is no longer separated to a function.

This is to allow for easier branching of behaviour as more commands are
supported. Future iterations can look into creating a separate chatbot
class and instance separate from the main() function.