Refactor Duke and TaskList
Duke:
- Remove all previous functions and replace with an echo() function.
- Standard messages are now stored as constants.

TaskList:
- markTask() and unmarkTask() now return String as output feedback.

Duke is refactored as many functions serve to only print a message.
It is redundant to create many functions to print to output.
echo() serves as a way to print output through the chatbot.

TaskList functions should return a message that will be sent as
feedback for the users. This is to make the chatbot feel more
responsive and users are aware of actions executed by chatbot.