Change functions to recognise date and time
Variables such as the deadline or start and end of events are stored as
Strings.

It makes more sense to store them as LocalDateTime for logical
validation.

Changed the declaration of the classes involved from String to
LocalDateTime.
Added functions in main class to convert between String and
LocalDateTime and replaced parts of the function to save and load from
the text file.