Add HardDisk.java to save the data from running Duke and reload previous data
Duke has an additional functionality of allowing data to be saved and reloaded.

HardDisk has a loadList and save function that allows data to be reloaded
and saved respectively.

Functions is seperated from the rest of the code within the harddisk class
as they have a common link in which they both utilise stored data