refactor+feat: Add todo,event,deadline + refactor
Add 3 variants of Tasks which all inherit from tasks. I also introduced OOP as I noticed a lot of duplicated code that's all in one file. I basically created a Parser that parses user input to ouptut commands, a State class to hold the state of the app, and variants of Commands to execute the user's intentions.