Save and load data in and from the hard disk
When the chatbot starts, it loads data from the hard disk. A folder named data and a file named harper.txt are created if not exist. When the task list changes, data are restored into the file. When the content of the file does not follow the expected format, the chatbot will terminate.