Enable saving of tasks in the hard disk
Implementing the feature of saving the task list into a text file in the
hard disk and also loading the task list from the respective file when
running the chatbot next time.

The format of the line shown when a task is stored has been changed to
ease the implementation of this feature.