feat: Add support for dates and times
The user can provide any string of characters as the date and time of
events and deadlines.

While this simplifies the parsing of user inputs and reduces the code
complexity, there is no checking done regarding the date time and
neither is there a standard format for the user

We add support for the parsing of the user input for the date and times
of events and deadlines, formatting them in the toString() method of
each task type model.

However, it currently relies on the user following a fixed format so as
to avoid DateTimeParseExceptions when parsing the user input which
should be fixed if we would like the user to have greater flexibility
and to reduce the number of false-positive exceptions occurring