Add LocalDate to Event
Event from and to was a String previously, giving no special meaning to it.

If from and to is a field that can capture date and have a temporal dimension we can do stuff related to time.
For example, we can give how much time remaining per task.

Change from, to from String to LocalDate
Add regex parsing to Event
Update Event saving and reading
Update Event creation command
Update help message for Event