Make a class for FileHandler and Command
The code for command is in the same file as main.

This makes it to maintain as the number of commands increases.

Let's put it in a separate file to fix this issue.

This makes it easier to maintian the code base.

A file for FileHandler is also created to handle the file system of the app.