Add Parser class for parsing string input and Command class and its subclasses
Command class does parsing pass the comaand to the Tasklist to handle it.

This is not consistent with the class definition and makes it hard to add new
commands to the app.

Let's create a new Parse class that does the parsing. Additionally, lets create
an abstract Command class that can run a function by itself and make all the
previous commands into individual classes that extends of the Command class.

This allows for better abstraction and makes the code more extensible.

Let's also add the Local Date time capability into the parser.