Modify the code base to be more OOP oriented
The code strucutre of Duke.java is adheres a lot more to functional programming.
With that said, the other parts of the code base follow OOP paradigm more.

As a result, there are inconsistencies in the main paradigm used by the code.

Let's restructure Duke.java to follow a more OOP approach. Additionally, let's rename
FileHandler to Storage and OutputMessage to Ui.

This can help to improve the consistency of the code base and increase its understandability.