Divide the classes into different packages
All the class files are in the same file, classes with varying behaviour
and type are mixed into the same place.

This makes the codebase hard to manage as it grows larger.

To solve this issue, let's divide the all the class files into different
packages, by grouping similar classes into the same package.

This way, the code base becomes clearer and more understandable.