Use Junit library to test a Parser, and a Storage method
The program is tested manually and tested with a semi-automatic CLI
testing method which incorporate input/output redirection technique.

Manual testing can take a lot of time and CLI testing is limited in that
it cannot test each method separately, introducing the risk of missing
some bugs.

Let's use Junit library to unit test a few of the non-trivial method. We
first start with testing the crucial, parse method of the Parser class,
and the non trivial method of saving the file with each item added into
the list.

Using Junit can help to isolate and locate bugs in individual methods
more effectively. Additionally it can help to automate the test process,
saving a lot of time.