Fix NullPointerException in empty component
There was a check to make sure that when no string without a / comes
after a component, it caused an exception. It did this by checking the
length, but the `StringBuilder` is null until after getting the first
part of the component parameter. Therefore, the length check caused a
NullPointerException.

The fix was to check if the builder is `null` instead.