Refactor code for better OOP design
The application is messy and scattered.

Makes it hard to maintain and extend.
Also makes it difficult to understand code.

Refactor application to adopt more OOP.
Creat new parser and dependencies to abstract
from the main application.

To improve code maintainability and reability,
allow for easy edits in future development.