Level-8
Updated Duke to recognize and parse date and time from input.

Added dateManager to handle parsing and output formatting of date and time.

Deadline and Event no longer holds their dates as a string, but as type "LocalDateTime".

Changed the structure of saving deadline and event types. Now it follows a fixed format of "dd MMM yyyy HH:mm".