A-MoreOOP
Updated Duke to be more OOP friendly.

Added taskManager to manage task related actions, exclusding fetching and storing of task data in csv file. That is still being maintained by fileManager.

Added commandManager to parse input and separate it into command word and description. It then directs the description to the appropriate managers and forwards the responses to Response class.

Added Response.java to handle the outputs of the Duke bot. In particular, it handles the reading of input and passing it to commandManager, then formats and outputs the responses appropriately to the interface.