Add save and load feature to Lindi
Lindi does not permanently store changes to the taskList, as it
is stored in a temporary array.

A new session would effectively wipe the users taskList. This is
not conducive for a user who wants to use the app for real life
task tracking.

Let's
* implement loading our current taskList when we start the session
* saving our taskList automatically whenever we make a change

We save it as a text in a txt file in a hidden folder in our current
directory '/.data/'. We save it as text for now because of ease of
implementation as well as it is human readable, which makes it easier
to debug.