Restructure code for better OOP design
The main Duke class have some common behaviours and repeated code with similar functionalities.

To enhance readability and maintainability, it's important to refactor the code to a more organized and modular OOP design.

Let's,
* Add a new Ui class that manages user interactions.
* Add a new Storage class that handles task file loading and saving
* Add a new Parser class that interprets user commands

This approach makes it easier to understand and maintai neach component. The classes encapsulate related functionality, allowing for easier maintenance in the future.