Add Level-7 requirements
Added loadData, appendTaskToFile, saveTaskListToFile, methods to allow
for saving and loading of the tasklist from the disk.

Added removeCorruptedData method to handle deletion of incorrect data at
the loadData stage. The method is designed to to only take an array of
indexes to allow for flexibility in what thrown exceptions are picked up
to be removed and to allow for deletion of larger numbers of corrupted
entries in a single pass.

Modified main method to use switch statement for todo, deadline and
event commands for better code readability.