Refactor code base into multiple packages
All classes are not in a package and are put in the java folder together. This makes it hard to distinguish related classes and makes the code base rather disorganized.

Let's refactor the code into three main packages:
* 'duke': Houses the main program and TaskList class.
* 'task': Manages the child tasks and the parent Task class.
* 'util': Handles user interactions.

This provides a clear separation of concerns and makes the code base organized and easy to navigate.