Add GUI to project
Gui version of the project has the same functionalities as the text-based ui.

Text-based ui can be difficult for the user to use as the output may be small.
Furthermore, text is in a single colour which is not exciting to the user.

Gui provides a more user-friendly platform for users to interact
with the program.

Let's apply Gui using a simpler method using JavaFx and gradle.