Package app as an executable JAR for easy distribution
Currently the app is not easy to distribute as it requires a lot of setup on the users' end.

Let's package the Duke project as a JAR so users can simply download and run the program.

Note that this commit does not have any significant change. (Removed extra line form TaskList)