Level-9
Added "find" feature to allow user to find tasks that matches the given substring.

Changed return type of "addTask", "deleteTask" and "setTaskStatus" in "taskManager" to Task (previously String). The formatting of output string will now be handled in commandManager.

Fixed bug that causes the fileWriter to throw an error when trying to write to file.