Duke Class: Update position of saveData() function in main()
The duke.txt file is supposed to be updated for every change made to
the list. Hence, the saveData() function has been moved to reflect this.
Prior to this, the .txt file would only be updated at the end of the
interaction with the Duke chat box when they use the command "bye".