Create new Package called Handler
I realised when implementing the parsing of duke.txt and loading the
contacts that the main method of Sir Duke was getting messy. With the
long block of switch-case, I decided to create a new package called the
handler, where I store functions to store how the parser is handled and
commands are going to be handled.