Add JUnit tests for Deadlines and Events
Add JUnit tests for Deadlines and Events to test the behavior of the respective classes.

There are cases where construction of a deadline or an event could result in a error due to the invalid parameters.

The addition of JUnit tests will allow for such bugs to be caught with greater ease.