Add Mark-as-done functionality
Added the ability to mark and unmark tasks in the list of tasks, and changed names of the tasklist ArrayList and the Task class. Added text output thatprints before a list is printed, in order to replcate the functionality in the example provided. As I had formerly implemented a Task class already, adding the mark-unmark functionality was trivial.