Add JUnit Tests to application
Testing is very manual, by printing out
variable and confirming each input by
hand.

This is not scalable and not practical
when the codebase becomes much bigger
and there are more dependencies in the
future.

Add JUnit tests for certain classes
based on importance. To ensure that the
expected behaviour is as observed and
does not change with further development.

JUnit tests are great as it is integrated
with gradle, thus becoming easy to maintain
and run.