Integrate GUI and Logic
Created a new class to replace the old UserInterface class. This is to
push functionality to the GraphicalUI class (formerly Graphics) and to
encapsulate all the logic in the Logic class.

There are still minor bug fixes to be made with the responses from the
chatbot, but the main functionality is there.

Tests were updated to reflect the change in the Task::toString method
as seen in a previous commit.

Duke class is now the main entry point of the application, and replaces
the Launcher class.