Package main application into JAR file
The application needs to be compiled and
ran with java from the terminal or IDE.

This will make the application difficult
more difficult to run for end users.

Package main application into java using
shadowjar, allow us to run the application
with a simple java - jar <filename>.

This will allow for an easier time to
test and use the application.