Add JavaDoc to application
Not all classes have completed Javadoc comments.

This makes reading the code confusing and time-
consuming as you have to understand the code before
making sense of it's purpose.

Add JavaDoc to each class and most methods.

This improves code readability and maintainability
as it allows readers to quickly understand what
the class and methods are used for.