Adhere to coding standard
The Java codes in the application do not
adhere to the coding standard.

This may result in the code being hard
to read or understand, and therefore even
harder to maintain

Change codes to adhere to coding standard,
imported checkstyler and used Gradle to
run.

Allow gradle to check Java code for
readability and makes code more readable.