Add find command to Chatbot application
There is no find command in the application,
users have to list all task and search manually.

This is no intuitive and causes the user to
manually search for his/her task.

Add find command to filter task for the user.

Allows user to quickly find the task they are
looking for.