Add toData() function
This function allows a different formatting that is used in the
text file, e.g. E | 1 | concert | Aug 2nd 4-6pm, as compared to
the toString method.

This format is also in accordance with how the text file will
be read when the programme runs.

Let's create a new function to give an output with a specific format