Create GUI for chatbot
There is no GUI for the chatbot and some logic can be better made to be OOP style.

With a GUI, it is easier for user to key in their commands. With more OOP, the code base is easier to read.

Use JavaFX to create the UI. Create Command and its corresponding subclasses to make the code more OOP.

JavaFX is utilised as it is incorporating well with Gradle. OOP is used to make the code more bearable to read.

Parser and TaskList Test cases were also modified to test the changed code in the respective classes.