Action and Task classes: add .equals()
Action and Task classes do not have custom equals() methods.

The equals() method is needed for assertEquals() to assert equality
between action and task objects based on the needs of this application.

Let's create equals() methods for all Action and Task classes and create
InputHandlerTest class to test for InputHandler class.