Add assertions
To ensure Duke is properly injected to MainWindow class.
To ensure that we are decoding the correct Task

Add an assertion on Duke that is not null when displaying initializing message
Add an assertion on Task to be an instance of event when decoding it.

Assertions help to identify potential bugs in programs.