Adjust sections based on feedback
Adjust sections of the code that were previously not as clear, change the way toString() method is handled for all task related classes and replace usage of concatenation in FileManager to be StringBuilder instead.

The goal is to be clearer in comments and error messages, extract common toString() code so the subclasses have fewer sections of repeated code and be more efficient in FileManager with the use of StringBuilder.

Thanks to chiralcentre for the useful feedback!