Add asserts
Program does not use any assertions, and have no validation of states
before running further operations.

This can cause the program to run with errors and cause unintended
consequences.

Assertions are added at various points of the script, mainly
Parser.java, Storage.java and Main.java.

This is done for early program termination in the case that important
assertions are not met.