Refactor ByeCommand message
ByeCommand returns a magic String literal. The String is then recognized
by MainWindow.java to know when to exit the program.

Typo or change in message can cause program to work incorrectly as it as
a functional component apart from printing to output.

Storing the message as a constant under ByeCommand will allow for better
consistency and improve readability.

In the future should avoid magic literals, especially if it affects
logic in the code.