Feat: add assertions
There are methods that require a a file in the file path to exist.

Checks need to be done to ensure the file location already exists to
prevent IO Exceptions.

Using assertions allows us to assume that file location already
exists at the point of saving or reading.

Let's assert that the file location exists during IO operations.