Mona: add assertions to methods
A switch statement in Mona currently does not use assertions.

Let's add assertions to the switch statement and the block after
to check against our assumption that currCommand can only take
the finite values that have been specified in the enumeration as well
as to ensure that execution in the getResponse() method does not
reach past the switch statement.

Assertions are statements to test assumptions in the program and
are used here in accordance to Java's specifications on programming
with Assertions to test for internal invariants.