Storage readLog(): Replace iterative implementation with Streams
readLog() utilises iteration to check through each line of the logs
and parses it into a relevant task before adding it to the task list.

Let's replace the iteration with the Streams API to pull the lines out
from the logs and map each line using relevant helper methods.

Java's Stream API represent a more functional programming way of
handling the same task which tends to be more descriptive for
readability and also encourages less mutability.