feat: Implement tagging feature for tasks
This commit introduces the ability for users to add tags to tasks using the format 'tag [task number] /tag [tag content]'. Tags are displayed alongside task descriptions and are prefixed with a '#'. The handleTag method has been updated to parse the command and append the tag to the specified task.