Add missing javadoc for getResponse method
The getResponse method helps to pass information from
the Numerator program to the UI. This method takes in
a String input and returns a String response to the
input.

Especially because this method will be used by other
packages, it is essential that the purpose of the
method is clear.

Adding javadoc helps to make the purpose clearer.