Add assertion to help validate filepath integrity
The load() attempts to read from the filepath if the
filepath points to a valid file. Otherwise, it
creates a new file at the filepath.

However, there may be existing regular files in the
filepath that may conflict with the creation of
the storage file. If the given filepath is invalid, the
application is unable to recursively modify the path
hierarchy to make the filepath valid.

Hence, the assertion is necessary to ensure filepath
validity.