Added assertions to the program
It is to document important assumptions that should hold at vaious points in the code. For example, the user entering text and pressing the enter button.

An AssertionError will be thrown and it will not be try and caught, indicating unrecoverable conditions. Let's handle edge cases in our program based on
the user's input text.