Support tentative scheduling for events
Events could only be scheduled to a specific date and time.

This commit adds the ability to schedule events tentatively, which means
that the event is not yet scheduled to a specific date and time.
The user is able to supply multiple tentative start and end times for
the event, and once a confirm command is issued, the tentative event
will become an event object with only one start and end time.

As a tentative event involves a data structure that is different from
the regular event, a new class TentativeEvent is created to handle
tentative events. An Event object is created from a TentativeEvent
object when the user confirms the tentative event.