Add CheckStyle to check for consistency
Codebase lacks the enforcement of proper
coding standards.

This could lead to messy, unformatted
code that makes the Codebase hard
to read and maintain.

Installed and configured a CheckStyle
xml file to run with gradlew, such
that every time I run the project,
it checks for the Java coding style.

This will ensure that before commiting
each version, I will have an
opportunity to check the Java style
before pushing to the repo.