Add Assertions
Added Java assertions to code
-start method in Ui class
The command variable might be null

Added assert to ensure that the command enumeration is not null before the switch statement.

This is done to prevent any NullPointerExceptions

-markTask method in TaskList class
The changes to the target Task might not be changed.

Added assert to ensure that the marked status of the target Task is equal to the intended boolean.

This is done to prevent any logical errors that might occur.

-unmarkTask method in TaskList class
The changes to the target Task might not be changed.

Added assert to ensure that the marked status of the target Task is equal to the intended boolean.

This is done to prevent any logical errors that might occur.