Comply method name to coding standard
There is a method that outputs the String description of the tasks to
the save file.

The peer review did point out that the method name can be confusing as
the class contains toString() method and the uses can be confusing for
readers.

As a step to make the code more readable, let's refactor all the method
name to a more explicit and clear name. Doing so will make the code much
more readable.