refactor: remove magic strings in the parser
The parser uses strings to match commands. This makes checking whether
the commands are valid to be error prone as a typo would mess it up. It
also makes it very inflexible as the string can only match one type.

To solution was to make an enum of the command type and moving command
type parsing logic to a static method in the enum.

This decouples checking the parsing to the enum, and it opens up to
more flexible parsing. This also means that we can write the parser and
be sure of exhaustiveness.