Add assertions to most methods
All methods all do not have assertion checks.

We should check for basic assumptions of some of the input, trivial
or not, so that runtime can verify them.

We add assertions to check whether input is not null for some methods,
we also check if some of the input equals a specific set of possible
input values. For example, action variable can either be
'mark' or 'unmark'.

Assertions are used as we can terminate the program if we fail to meet
basic assumptions during run-time. These are different from expected
exceptions.