Abstract the pattern regex from Storage class to adhere to SLAP
Currently, storage class has no abstraction for the pattern regex
and implementation to check for regex before adding to list of tasks
is present alongside other abstraction.

This violates SLAP and the method is simply too long.

We abstract the pattern regex to PatternParser as earlier commits did.

We adhere to OOP and make use of an already useful class that is made
for such pattern regex abstraction.