Extensions: Add method to update task name of any type, and method to check for datetime clashes for Event type tasks when adding new events
For the Update extension, it is necessary for user to be able to rectify mistakes in their tasks or simply update the descriptions if items change. For the Detecting Clashes extension, it is necessary for Nicole to be a coherent task manager and disallow obvious clashes in schedule.

For Update, rather than making users need to delete items to re-add them with the desired changes, we can allow them to change items directly

The Update extension requires a change of how task name is stored for Deadline and Event. Instead of storing the whole description including datetime, now only the description is stored and datetime is concatenated to their toString separately