Fixed bug where welcome message is not displayed
Upon startup, Blu does not display welcome message as expected.
This is because the method to display the welcome message is
never called in the application.

Create the welcome message in init() and displays it in start().
Change welcome message to include canonical path of the storage
file instead.

This allows users to locate the file path easily.