Use Assertions in the Command Package
Modified the gradle setting to enable Java' assertions at runtime

Assertions have been added to most of the Command classes to help
detect whether the type index conforms to the set optional range.

Add assertions after processing the input length to determine if
the input still does not conform to the expected specification.