Add update feature
Currently, there is no update feature. We are forced to delete and
reinput the task if needed.

This is clearly inconvenient as reinputting is not easy.

Now we can select a index and update its appropriate field without
changing other fields. We do so by simple update command.

Since our constraint is CLI and single line is preferred over
multi lines, we utilize a simple single line update command.

Tests still have to be added however.