Add assertions to non-empty command and file write
Program assumes command parse by input parser is always a
non-null value. It also assumes that the tasks will be
successfully saved to the file.

This can lead to unintended bugs caused by the developer,
as both cases are not supposed to be empty under any
circumstances.

Use assert statements to indicate a possible bug within
these two cases, which violates the assumption that both
commands and file write operations should be non-empty.

This is allow runtime to detect any bugs made by the developer,
allowing the program to terminate earlier safely.